+++ browser/data-browser/src/views/Installation/InstallationPage.tsx
@@ -361,6 +361,7 @@ export const InstallationPage: React.FC<
             canWrite={canWrite}
             platforms={proxyPlatforms}
             connected={connected}
+            pluginName={name || title || resource.subject}
           />
         )}
         {pluginAgent && (